+++ app/models/variable.rb
@@ -99,7 +99,7 @@ class Variable < ActiveRecord::Base
   end
 
   def editable_by?(current_user)
-    current_user.all_variables.pluck(:id).include?(self.id)
+    current_user.all_variables.where(id: self.id).count == 1
   end
 
   def copyable_attributes